backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chatgpt-backend"
version = "0.1.0"
description = "ChatGPT-like 对话系统后端"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]
